Extract the queue and album-detail dialogs
Pull the queue dialog and the album-detail dialog out of MainWindow. The queue's
drag-reorder math becomes a unit-tested QueueReorderModel; album detail keeps its
release picker and track list but hands each per-release action (export, delete,
edit, re-identify, gallery, change cover) to a ReleaseActionDialogs class, and
the per-track export flow becomes a method on the album-detail dialog. The
export-selection constructors and value equality — shared by album detail and
the settings export section — move to a single ExportSelection helper.
