+++ bae-windows.Tests/bae-windows.Tests.csproj
@@ -52,6 +52,7 @@
     <Compile Include="../bae-windows/UpdateFlow.cs" Link="UpdateFlow.cs" />
     <Compile Include="../bae-windows/Stores/SortState.cs" Link="SortState.cs" />
     <Compile Include="../bae-windows/TextTruncation.cs" Link="TextTruncation.cs" />
+    <Compile Include="../bae-windows/Stores/QueueReorderModel.cs" Link="QueueReorderModel.cs" />
   </ItemGroup>
 
   <ItemGroup>
