+++ bae-windows/MainWindow.xaml.cs
@@ -28,13 +28,6 @@ namespace Bae.Windows;
 /// </summary>
 public sealed partial class MainWindow : Window
 {
-    private sealed record QueueEntryRow(BridgeQueueEntry Entry)
-    {
-        internal string EntryId => Entry.EntryId;
-        public override string ToString() =>
-            $"{Entry.Title} — {Entry.ArtistNames} · {Loc.Duration(Entry.DurationMs)}".Trim();
-    }
-
     private bool _populatingSortBox;
 
     // The library session (handle + event subscription) and the stores it drives.
@@ -57,6 +50,9 @@ public sealed partial class MainWindow : Window
     private readonly ImportDialog _importDialog;
     private readonly ImportPickerDialog _importPicker;
     private readonly ImportConfirmDialog _importConfirm;
+    private readonly ReleaseActionDialogs _releaseActions;
+    private readonly AlbumDetailDialog _albumDetail;
+    private readonly QueueDialog _queueDialog;
 
     // Releases whose unmanage is running right now. Unmanage is a blocking
     // foreground transfer (unlike pin, which enqueues, or upload, which lives in
@@ -164,6 +160,22 @@ public sealed partial class MainWindow : Window
         _router = new UiEventRouter(_playback, _shell, _projections, _mediaControls, _import.HandlePreviewEvent);
         _session.UiEvent += _router.Route;
 
+        // Album detail and the per-release action dialogs it opens, plus the queue
+        // dialog. Album detail is the shared entry point from the grid, the panes,
+        // the now-playing jump, and the import "view in library" banner.
+        _releaseActions = new ReleaseActionDialogs(
+            _session,
+            () => Content.XamlRoot,
+            () => WinRT.Interop.WindowNative.GetWindowHandle(this),
+            text => StatusText.Text = text);
+        _albumDetail = new AlbumDetailDialog(
+            _session,
+            () => Content.XamlRoot,
+            () => WinRT.Interop.WindowNative.GetWindowHandle(this),
+            text => StatusText.Text = text,
+            _releaseActions);
+        _queueDialog = new QueueDialog(_session, _playback, () => Content.XamlRoot);
+
         // The composer/search panes and the library-lifecycle dialogs. The window
         // stays the navigation shell (open/close/switch); these render and drive
         // their own screens, calling back for the operations it owns.
@@ -174,7 +186,7 @@ public sealed partial class MainWindow : Window
             ComposerDetailPane,
             text => StatusText.Text = text,
             ShowComposerBrowser,
-            (albumId, scrollToTrackId, initialReleaseId) => ShowAlbumDetail(albumId, scrollToTrackId, initialReleaseId));
+            _albumDetail.Show);
         _unlockDialog = new UnlockDialog(() => Content.XamlRoot, text => StatusText.Text = text, OpenLibrary);
         _joinDialog = new JoinLibraryDialog(
             () => Content.XamlRoot,
@@ -207,7 +219,7 @@ public sealed partial class MainWindow : Window
         // The import flow: the confirm step (which can open an album), the picker
         // that leads to it, and the folder-scan dialog that opens the picker.
         _importConfirm = new ImportConfirmDialog(
-            _session, () => Content.XamlRoot, albumId => ShowAlbumDetail(albumId));
+            _session, () => Content.XamlRoot, albumId => _albumDetail.Show(albumId));
         _importPicker = new ImportPickerDialog(_session, () => Content.XamlRoot, _import, _importConfirm);
         _importDialog = new ImportDialog(
             _session,
@@ -655,231 +667,9 @@ public sealed partial class MainWindow : Window
         _shell.ShowBanner(InfoBarSeverity.Error, Loc.Chrome("import.error_title"), message);
     }
 
-    /// <summary>
-    /// A dismiss-only error dialog: a title, an optional detail body, and a single
-    /// "OK" button that closes it.
-    /// </summary>
-    private async System.Threading.Tasks.Task ShowError(string title, string? message = null)
-    {
-        var dialog = new ContentDialog
-        {
-            Title = title,
-            CloseButtonText = Loc.Chrome("action.ok"),
-            XamlRoot = Content.XamlRoot,
-        };
-        if (message is not null)
-        {
-            dialog.Content = message;
-        }
-
-        await dialog.ShowAsync();
-    }
-
     private async void OnQueueClick(object sender, RoutedEventArgs e)
     {
-        if (CurrentHandleOrNull() == null)
-        {
-            return;
-        }
-
-        // Clear empties only the manual lane (the context survives), so it
-        // disables on an empty manual lane regardless of the context.
-        var clear = new Button
-        {
-            Content = Loc.Chrome("queue.clear"),
-            IsEnabled = _playback.ManualQueue.Count > 0,
-        };
-        clear.Click += (_, _) => WithCurrentHandle(NativeBae.QueueClear);
-
-        var content = new StackPanel { Spacing = 8 };
-        content.Children.Add(clear);
-
-        // Two distinct sections: the manual lane ("Up Next"), then the context
-        // (the release being played from). Each is its own reorderable list —
-        // entry ids are unique across both and core no-ops a cross-lane move.
-        if (_playback.ManualQueue.Count > 0)
-        {
-            content.Children.Add(QueueSectionLabel(Loc.Chrome("queue.section.up_next")));
-            content.Children.Add(BuildQueueLaneList(_playback.ManualQueue));
-        }
-
-        if (_playback.Context is { Upcoming.Length: > 0 } ctx)
-        {
-            // The context section names what's playing — a release ("Playing From")
-            // vs the whole library — by the source kind the wire shape carries.
-            var labelKey = ctx.Kind == BridgePlaybackSourceKind.Library
-                ? "queue.section.your_library"
-                : "queue.section.playing_from";
-            content.Children.Add(ContextSectionLabel(Loc.Chrome(labelKey), ctx.Shuffled));
-            content.Children.Add(BuildQueueLaneList(ctx.Upcoming));
-        }
-
-        var dialog = new ContentDialog
-        {
-            Title = Loc.Chrome("queue.title"),
-            Content = content,
-            CloseButtonText = Loc.Chrome("action.close"),
-            XamlRoot = Content.XamlRoot,
-        };
-        await dialog.ShowAsync();
-    }
-
-    // A plain section header for the queue dialog (the manual "Up Next" lane,
-    // which is never shuffled and has no shuffle control).
-    private static TextBlock QueueSectionLabel(string text) => new()
-    {
-        Text = text,
-        FontWeight = Microsoft.UI.Text.FontWeights.SemiBold,
-    };
-
-    // The context section's header, with a shuffle toggle that flips the context
-    // between sequential and shuffled order while the current track keeps playing.
-    private StackPanel ContextSectionLabel(string text, bool shuffled)
-    {
-        var row = new StackPanel
-        {
-            Orientation = Orientation.Horizontal,
-            Spacing = 6,
-            VerticalAlignment = VerticalAlignment.Center,
-        };
-        row.Children.Add(new TextBlock
-        {
-            Text = text,
-            FontWeight = Microsoft.UI.Text.FontWeights.SemiBold,
-            VerticalAlignment = VerticalAlignment.Center,
-        });
-        // Segoe MDL2 Assets "Shuffle" glyph (U+E8B1), accented when on.
-        var toggle = new Button
-        {
-            Content = new FontIcon
-            {
-                Glyph = "\uE8B1",
-                FontSize = 14,
-                Foreground = shuffled
-                    ? (Brush)Application.Current.Resources["AccentTextFillColorPrimaryBrush"]
-                    : (Brush)Application.Current.Resources["TextFillColorSecondaryBrush"],
-            },
-            Padding = new Thickness(6, 2, 6, 2),
-        };
-        Microsoft.UI.Xaml.Automation.AutomationProperties.SetName(
-            toggle, Loc.Chrome(shuffled ? "queue.shuffle.off" : "queue.shuffle.on"));
-        ToolTipService.SetToolTip(
-            toggle, Loc.Chrome(shuffled ? "queue.shuffle.off" : "queue.shuffle.on"));
-        toggle.Click += (_, _) => WithCurrentHandle(
-            handle => NativeBae.SetShuffle(handle, !shuffled));
-        row.Children.Add(toggle);
-        return row;
-    }
-
-    // One lane's reorderable list: click skips, right-tap removes, drag reorders
-    // within the lane (the framework raises a Move, forwarded to core by entry id).
-    private ListView BuildQueueLaneList(IEnumerable<BridgeQueueEntry> items)
-    {
-        var queueItems = new ObservableCollection<QueueEntryRow>(items.Select(item => new QueueEntryRow(item)));
-        queueItems.CollectionChanged += (_, args) =>
-        {
-            if (args.Action == System.Collections.Specialized.NotifyCollectionChangedAction.Move)
-            {
-                // The collection already reflects the move: the entry now sits at
-                // NewStartingIndex and lands before whatever follows it (null when
-                // it's now last = move to the lane's end).
-                var moved = queueItems[args.NewStartingIndex];
-                var beforeIndex = args.NewStartingIndex + 1;
-                var beforeEntryId = beforeIndex < queueItems.Count
-                    ? queueItems[beforeIndex].EntryId
-                    : null;
-                WithCurrentHandle(
-                    handle => NativeBae.QueueReorder(handle, moved.EntryId, beforeEntryId));
-            }
-        };
-
-        var list = new ListView
-        {
-            ItemsSource = queueItems,
-            SelectionMode = ListViewSelectionMode.None,
-            IsItemClickEnabled = true,
-            CanReorderItems = true,
-            CanDragItems = true,
-            AllowDrop = true,
-        };
-        list.ItemClick += (_, args) =>
-        {
-            if (args.ClickedItem is QueueEntryRow clicked)
-            {
-                WithCurrentHandle(handle => NativeBae.QueueSkipTo(handle, clicked.EntryId));
-            }
-        };
-        // Right-tap a row to drop it from the queue. Removing locally too keeps the
-        // open dialog in sync; the Move-only reorder handler ignores this Remove.
-        list.RightTapped += (_, args) =>
-        {
-            if (args.OriginalSource is not FrameworkElement element
-                || element.DataContext is not QueueEntryRow item)
-            {
-                return;
-            }
-
-            var index = queueItems.IndexOf(item);
-            if (index < 0)
-            {
-                return;
-            }
-
-            var menu = new MenuFlyout();
-            var remove = new MenuFlyoutItem { Text = Loc.Chrome("queue.remove_item") };
-            remove.Click += (_, _) =>
-            {
-                // The generated bridge removes by entry id, so a reorder between the right-tap and
-                // the click can't target the wrong row. The local index is only to keep
-                // the open dialog's collection in sync.
-                var idx = queueItems.IndexOf(item);
-                if (idx < 0)
-                {
-                    return;
-                }
-                WithCurrentHandle(handle => NativeBae.QueueRemove(handle, item.EntryId));
-                queueItems.RemoveAt(idx);
-            };
-            menu.Items.Add(remove);
-            menu.ShowAt(element, new FlyoutShowOptions { Position = args.GetPosition(element) });
-        };
-
-        return list;
-    }
-
-    /// <summary>
-    /// A cover-art thumbnail tile: an image over a one-line caption, the whole
-    /// tile a borderless button. The caller wires <c>Click</c> — the
-    /// change-cover gallery applies the selection immediately, the import
-    /// confirm gallery carries it and highlights the picked tile via the
-    /// button's border.
-    /// </summary>
-    private static Button CoverTile(ImageSource? source, string caption)
-    {
-        var thumb = new Image
-        {
-            Source = source,
-            Stretch = Stretch.UniformToFill,
-            Width = 120,
-            Height = 120,
-        };
-        var label = new TextBlock
-        {
-            Text = caption,
-            TextTrimming = TextTrimming.CharacterEllipsis,
-            MaxWidth = 120,
-            HorizontalAlignment = HorizontalAlignment.Center,
-        };
-        var stack = new StackPanel { Spacing = 4 };
-        stack.Children.Add(thumb);
-        stack.Children.Add(label);
-        return new Button
-        {
-            Content = stack,
-            Background = new SolidColorBrush(Microsoft.UI.Colors.Transparent),
-            BorderThickness = new Thickness(0),
-            Padding = new Thickness(4),
-        };
+        await _queueDialog.Show();
     }
 
     private void OnPlayPause(object sender, RoutedEventArgs e)
@@ -940,7 +730,7 @@ public sealed partial class MainWindow : Window
             return;
         }
 
-        await ShowAlbumDetail(albumId, scrollToTrackId: _playback.NowPlayingTrackId);
+        await _albumDetail.Show(albumId, scrollToTrackId: _playback.NowPlayingTrackId);
     }
 
     // Space toggles play/pause from anywhere — except while typing in a text
@@ -1023,1052 +813,7 @@ public sealed partial class MainWindow : Window
             return;
         }
 
-        await ShowAlbumDetail(album.Id);
-    }
-
-    /// <summary>
-    /// Open the album-detail dialog for <paramref name="albumId"/>: the album's
-    /// releases (with a picker when there's more than one), the track list, and
-    /// the per-release actions (play / queue / edit / re-identify / gallery /
-    /// change cover / delete). Reused by the album grid (<see cref="OnAlbumClick"/>)
-    /// and the import confirmation's "view in library" banner.
-    /// </summary>
-    private async System.Threading.Tasks.Task ShowAlbumDetail(
-        string albumId,
-        string? scrollToTrackId = null,
-        string? initialReleaseId = null)
-    {
-        var (current, response) = await RunForCurrentHandle(
-            handle => NativeBae.GetAlbumDetail(handle, albumId));
-        if (!current)
-        {
-            return;
-        }
-        if (response.Error is not null || response.Detail is null)
-        {
-            StatusText.Text = response.Error ?? Loc.Chrome("album.open_failed");
-            return;
-        }
-
-        var detail = response.Detail;
-
-        if (detail.Releases.Count == 0)
-        {
-            StatusText.Text = Loc.Chrome("album.open_failed");
-            return;
-        }
-
-        // The release the dialog acts on: the user's primary, or the first. The
-        // picker (added below when there's more than one) reassigns it, and every
-        // per-release action and the track list read it, so switching release
-        // retargets play / queue / edit / gallery / delete to that release.
-        // When revealing a now-playing track, start on the release that actually
-        // contains it — which may not be the primary.
-        Release? trackRelease = scrollToTrackId is null
-            ? null
-            : detail.Releases.FirstOrDefault(r => r.Tracks.Any(t => t.TrackId == scrollToTrackId));
-        Release selectedRelease = trackRelease
-            ?? detail.Releases.FirstOrDefault(r => r.ReleaseId == initialReleaseId)
-            ?? detail.Releases.FirstOrDefault(r => r.ReleaseId == detail.PrimaryReleaseId)
-            ?? detail.Releases[0];
-
-        var header = new StackPanel
-        {
-            Orientation = Orientation.Horizontal,
-            Spacing = 12,
-        };
-        header.Children.Add(new TextBlock
-        {
-            Text = detail.Artist,
-            VerticalAlignment = VerticalAlignment.Center,
-        });
-        var shuffleButton = new Button { Content = Loc.Chrome("album.shuffle") };
-        var editButton = new Button { Content = Loc.Chrome("album.edit.label") };
-        var reidentifyButton = new Button { Content = Loc.Chrome("album.reidentify.label") };
-        header.Children.Add(shuffleButton);
-        header.Children.Add(editButton);
-        header.Children.Add(reidentifyButton);
-
-        // Overflow menu: queueing + delete (release-level, so they need no track ids).
-        var deleteRequested = false;
-        var moreButton = new Button { Content = "⋯" };
-        var moreMenu = new MenuFlyout();
-        var playNextItem = new MenuFlyoutItem { Text = Loc.Chrome("menu.play_next") };
-        playNextItem.Click += (_, _) =>
-        {
-            if (!string.IsNullOrEmpty(selectedRelease.ReleaseId))
-            {
-                WithCurrentHandle(
-                    handle => NativeBae.AddReleaseNext(handle, selectedRelease.ReleaseId));
-            }
-        };
-        var addQueueItem = new MenuFlyoutItem { Text = Loc.Chrome("menu.add_to_queue") };
-        addQueueItem.Click += (_, _) =>
-        {
-            if (!string.IsNullOrEmpty(selectedRelease.ReleaseId))
-            {
-                WithCurrentHandle(
-                    handle => NativeBae.AddReleaseToQueue(handle, selectedRelease.ReleaseId));
-            }
-        };
-        // Set the selected release as the album's primary (canonical) one — only
-        // meaningful when the album has more than one release.
-        var setPrimaryItem = new MenuFlyoutItem { Text = Loc.Chrome("menu.set_primary") };
-        setPrimaryItem.Click += async (_, _) =>
-        {
-            var (current, error) = await RunForCurrentHandle(
-                handle => NativeBae.SetPrimaryRelease(
-                    handle,
-                    detail.Id,
-                    selectedRelease.ReleaseId));
-            if (!current)
-            {
-                return;
-            }
-            StatusText.Text = error ?? Loc.Chrome("menu.set_primary_done");
-        };
-        var changeCoverItem = new MenuFlyoutItem { Text = Loc.Chrome("menu.change_cover") };
-        var exportReleaseItem = new MenuFlyoutItem { Text = Loc.Chrome("menu.export") };
-        var deleteItem = new MenuFlyoutItem { Text = Loc.Chrome("menu.delete") };
-        moreMenu.Items.Add(playNextItem);
-        moreMenu.Items.Add(addQueueItem);
-        if (detail.Releases.Count > 1)
-        {
-            moreMenu.Items.Add(setPrimaryItem);
-        }
-        moreMenu.Items.Add(exportReleaseItem);
-        moreMenu.Items.Add(changeCoverItem);
-        moreMenu.Items.Add(deleteItem);
-        moreButton.Flyout = moreMenu;
-        header.Children.Add(moreButton);
-
-        // Export failures surface here, inside the dialog: the window-level banner
-        // is occluded by this modal album-detail dialog, so a banner error would be
-        // invisible until the dialog is dismissed.
-        var exportStatus = new TextBlock
-        {
-            TextWrapping = TextWrapping.Wrap,
-            Visibility = Visibility.Collapsed,
-        };
-
-        // Track list: click a row to play the release from that track; right-tap
-        // for per-track queueing. The play index is the track's position in the
-        // release's track list, which is what PlayRelease expects.
-        var trackList = new ListView
-        {
-            ItemsSource = selectedRelease.Tracks,
-            SelectionMode = ListViewSelectionMode.None,
-            IsItemClickEnabled = true,
-        };
-        // "Go to now playing" reveal: once the list is realized, scroll the target
-        // track into view and flash it. selectedRelease was chosen to contain it.
-        if (scrollToTrackId is not null
-            && selectedRelease.Tracks.FirstOrDefault(t => t.TrackId == scrollToTrackId) is { } revealTrack)
-        {
-            trackList.Loaded += (_, _) =>
-            {
-                trackList.ScrollIntoView(revealTrack);
-                FlashTrackRowWhenRealized(trackList, revealTrack, attemptsLeft: 8);
-            };
-        }
-        void PlayFromTrack(Track track)
-        {
-            if (!string.IsNullOrEmpty(selectedRelease.ReleaseId))
-            {
-                WithCurrentHandle(
-                    handle => NativeBae.PlayRelease(
-                        handle, selectedRelease.ReleaseId, selectedRelease.Tracks.IndexOf(track), false));
-            }
-        }
-        void QueueTrack(Track track, bool next)
-        {
-            var (current, error) = WithCurrentHandle(handle => next
-                ? NativeBae.AddNext(handle, new[] { track.TrackId })
-                : NativeBae.AddToQueue(handle, new[] { track.TrackId }));
-            if (!current)
-            {
-                return;
-            }
-            if (error is not null)
-            {
-                StatusText.Text = error;
-            }
-        }
-        trackList.ItemClick += (_, args) =>
-        {
-            if (args.ClickedItem is Track track)
-            {
-                PlayFromTrack(track);
-            }
-        };
-        trackList.RightTapped += (_, args) =>
-        {
-            if (args.OriginalSource is not FrameworkElement element || element.DataContext is not Track track)
-            {
-                return;
-            }
-
-            var menu = new MenuFlyout();
-            var play = new MenuFlyoutItem { Text = Loc.Chrome("menu.play") };
-            play.Click += (_, _) => PlayFromTrack(track);
-            var playNextTrack = new MenuFlyoutItem { Text = Loc.Chrome("menu.play_next") };
-            playNextTrack.Click += (_, _) => QueueTrack(track, next: true);
-            var addQueueTrack = new MenuFlyoutItem { Text = Loc.Chrome("menu.add_to_queue") };
-            addQueueTrack.Click += (_, _) => QueueTrack(track, next: false);
-            var exportTrack = new MenuFlyoutItem { Text = Loc.Chrome("menu.export") };
-            exportTrack.Click += async (_, _) =>
-            {
-                exportStatus.Visibility = Visibility.Collapsed;
-                var picker = new global::Windows.Storage.Pickers.FileSavePicker();
-                WinRT.Interop.InitializeWithWindow.Initialize(
-                    picker, WinRT.Interop.WindowNative.GetWindowHandle(this));
-                var (settingsCurrent, settings) = await RunForCurrentHandle(NativeBae.GetSettings);
-                if (!settingsCurrent)
-                {
-                    return;
-                }
-                var trackPresets = settings.ExportPresets
-                    .Where(preset => preset.AppliesToTrack)
-                    .ToList();
-                var originalSelection = OriginalExportSelection();
-                var (extensionCurrent, originalExtension) = await RunForCurrentHandle(
-                    handle => NativeBae.ExportTrackExtension(handle, track.TrackId, originalSelection));
-                if (!extensionCurrent)
-                {
-                    return;
-                }
-                if (originalExtension is null)
-                {
-                    exportStatus.Text = Loc.Chrome("track.export.prepare_failed");
-                    exportStatus.Visibility = Visibility.Visible;
-                    return;
-                }
-                var choices = new List<(string Label, string Extension, BridgeExportSelection Selection)>
-                {
-                    (Loc.Chrome("track.export.original"), $".{originalExtension}", originalSelection),
-                };
-                choices.AddRange(trackPresets.Select(preset => (
-                    preset.TrackPickerLabel,
-                    preset.FileExtension,
-                    PresetExportSelection(preset.Id))));
-                var formatPicker = new ComboBox
-                {
-                    Header = Loc.Chrome("settings.export.default_track_format"),
-                    MinWidth = 260,
-                };
-                var defaultIndex = -1;
-                for (var index = 0; index < choices.Count; index++)
-                {
-                    var choice = choices[index];
-                    if (ExportSelectionsEqual(choice.Selection, settings.DefaultTrackExportSelection))
-                    {
-                        defaultIndex = index;
-                    }
-                    formatPicker.Items.Add(new ComboBoxItem
-                    {
-                        Content = choice.Label,
-                    });
-                }
-                if (defaultIndex < 0)
-                {
-                    exportStatus.Text = Loc.Chrome("track.export.prepare_failed");
-                    exportStatus.Visibility = Visibility.Visible;
-                    return;
-                }
-                formatPicker.SelectedIndex = defaultIndex;
-                var formatDialog = new ContentDialog
-                {
-                    Title = Loc.Chrome("settings.export.label"),
-                    Content = formatPicker,
-                    PrimaryButtonText = Loc.Chrome("settings.export.label"),
-                    CloseButtonText = Loc.Chrome("action.cancel"),
-                    XamlRoot = Content.XamlRoot,
-                };
-                var formatResult = await formatDialog.ShowAsync();
-                if (formatResult != ContentDialogResult.Primary)
-                {
-                    return;
-                }
-                if (formatPicker.SelectedIndex < 0)
-                {
-                    exportStatus.Text = Loc.Chrome("track.export.prepare_failed");
-                    exportStatus.Visibility = Visibility.Visible;
-                    return;
-                }
-                var selectedChoice = choices[formatPicker.SelectedIndex];
-                picker.FileTypeChoices.Add(
-                    selectedChoice.Label,
-                    new List<string> { selectedChoice.Extension });
-                // Seed the suggested name from the configured filename template,
-                // which the core renders and sanitizes from this track's metadata
-                // (falling back to the title, then a fixed stem, on its own). A
-                // null return — or a throw — means that render failed (the core
-                // logged the cause); surface it and abort rather than exporting
-                // under a guessed name.
-                string? stem;
-                try
-                {
-                    var (nameCurrent, suggestedName) = await RunForCurrentHandle(
-                        handle => NativeBae.ExportTrackSuggestedName(handle, track.TrackId));
-                    if (!nameCurrent)
-                    {
-                        return;
-                    }
-                    stem = suggestedName;
-                }
-                catch (Exception ex)
-                {
-                    BaeDiagnostics.Logger.Error("export suggested-name lookup threw", ex);
-                    stem = null;
-                }
-                if (stem is null)
-                {
-                    exportStatus.Text = Loc.Chrome("track.export.prepare_failed");
-                    exportStatus.Visibility = Visibility.Visible;
-                    return;
-                }
-                picker.SuggestedFileName = stem;
-                var file = await picker.PickSaveFileAsync();
-                if (file is null)
-                {
-                    return;
-                }
-
-                var path = file.Path;
-                var (exportCurrent, error) = await RunForCurrentHandle(
-                    handle => NativeBae.ExportTrack(handle, track.TrackId, path, selectedChoice.Selection));
-                if (!exportCurrent)
-                {
-                    return;
-                }
-                if (error is not null)
-                {
-                    exportStatus.Text = error;
-                    exportStatus.Visibility = Visibility.Visible;
-                }
-            };
-            menu.Items.Add(play);
-            menu.Items.Add(playNextTrack);
-            menu.Items.Add(addQueueTrack);
-            menu.Items.Add(exportTrack);
-            menu.ShowAt(element, new FlyoutShowOptions { Position = args.GetPosition(element) });
-        };
-
-        var content = new StackPanel { Spacing = 8 };
-        content.Children.Add(header);
-        // Release picker, only when the album has more than one pressing. Choosing
-        // a release swaps the track list and retargets every per-release action.
-        if (detail.Releases.Count > 1)
-        {
-            var releasePicker = new ComboBox
-            {
-                Header = Loc.Chrome("album.release_picker"),
-                ItemsSource = detail.Releases,
-                SelectedItem = selectedRelease,
-            };
-            releasePicker.SelectionChanged += (_, _) =>
-            {
-                if (releasePicker.SelectedItem is Release release)
-                {
-                    selectedRelease = release;
-                    trackList.ItemsSource = selectedRelease.Tracks;
-                }
-            };
-            content.Children.Add(releasePicker);
-        }
-        content.Children.Add(trackList);
-        content.Children.Add(exportStatus);
-
-        var dialog = new ContentDialog
-        {
-            Title = detail.Title,
-            Content = content,
-            PrimaryButtonText = Loc.Chrome("action.play"),
-            SecondaryButtonText = Loc.Chrome("album.gallery"),
-            CloseButtonText = Loc.Chrome("action.close"),
-            XamlRoot = Content.XamlRoot,
-        };
-
-        // Edit and re-identify each replace the detail dialog: close it, then open
-        // the chosen sheet on the primary release. ShowAsync resolves with None
-        // when Hide() is called.
-        var editRequested = false;
-        var reidentifyRequested = false;
-        var shuffleRequested = false;
-        editButton.Click += (_, _) =>
-        {
-            editRequested = true;
-            dialog.Hide();
-        };
-        reidentifyButton.Click += (_, _) =>
-        {
-            reidentifyRequested = true;
-            dialog.Hide();
-        };
-        shuffleButton.Click += (_, _) =>
-        {
-            shuffleRequested = true;
-            dialog.Hide();
-        };
-        // Changing the cover opens its own gallery dialog; a nested ContentDialog
-        // can't open over this one, so close it first and open the gallery after
-        // ShowAsync returns (the gallery/edit/re-identify pattern).
-        var changeCoverRequested = false;
-        var exportReleaseRequested = false;
-        changeCoverItem.Click += (_, _) =>
-        {
-            changeCoverRequested = true;
-            dialog.Hide();
-        };
-        exportReleaseItem.Click += (_, _) =>
-        {
-            exportReleaseRequested = true;
-            dialog.Hide();
-        };
-        deleteItem.Click += (_, _) =>
-        {
-            deleteRequested = true;
-            dialog.Hide();
-        };
-
-        var result = await dialog.ShowAsync();
-        if (editRequested)
-        {
-            await ShowEditMetadata(selectedRelease.ReleaseId);
-        }
-        else if (reidentifyRequested)
-        {
-            await ShowReidentify(selectedRelease.ReleaseId, detail.Artist, detail.Title);
-        }
-        else if (shuffleRequested)
-        {
-            if (!string.IsNullOrEmpty(selectedRelease.ReleaseId))
-            {
-                WithCurrentHandle(
-                    handle => NativeBae.PlayRelease(handle, selectedRelease.ReleaseId, -1, true));
-            }
-        }
-        else if (changeCoverRequested)
-        {
-            await ShowChangeCover(detail.Id, selectedRelease.ReleaseId);
-        }
-        else if (exportReleaseRequested && !string.IsNullOrEmpty(selectedRelease.ReleaseId))
-        {
-            await ShowExportRelease(selectedRelease.ReleaseId);
-        }
-        else if (deleteRequested)
-        {
-            await ConfirmDeleteRelease(selectedRelease.ReleaseId);
-        }
-        else if (result == ContentDialogResult.Primary && !string.IsNullOrEmpty(selectedRelease.ReleaseId))
-        {
-            WithCurrentHandle(
-                handle => NativeBae.PlayRelease(handle, selectedRelease.ReleaseId, -1, false));
-        }
-        else if (result == ContentDialogResult.Secondary)
-        {
-            await ShowGallery(selectedRelease.ReleaseId);
-        }
-    }
-
-    // ScrollIntoView realizes the target row's container only after a later layout
-    // pass, so poll for it across a few UI ticks before flashing. Without this the
-    // flash silently no-ops for any track below the initial viewport — the common
-    // case for "go to now playing". Gives up after a bounded number of attempts.
-    private static void FlashTrackRowWhenRealized(ListView list, Track track, int attemptsLeft)
-    {
-        if (list.ContainerFromItem(track) is ListViewItem row)
-        {
-            FlashRow(row);
-            return;
-        }
-
-        if (attemptsLeft <= 0)
-        {
-            return;
-        }
-
-        list.DispatcherQueue.TryEnqueue(() => FlashTrackRowWhenRealized(list, track, attemptsLeft - 1));
-    }
-
-    // Tint a row with the system accent and fade it out over three seconds — the
-    // "go to now playing" flash, mirroring macOS.
-    private static void FlashRow(ListViewItem row)
-    {
-        var accent = new global::Windows.UI.ViewManagement.UISettings()
-            .GetColorValue(global::Windows.UI.ViewManagement.UIColorType.Accent);
-        var brush = new SolidColorBrush(accent) { Opacity = 0.35 };
-        row.Background = brush;
-
-        var fade = new DoubleAnimation
-        {
-            To = 0,
-            Duration = new Duration(TimeSpan.FromSeconds(3)),
-            EnableDependentAnimation = true,
-        };
-        Storyboard.SetTarget(fade, brush);
-        Storyboard.SetTargetProperty(fade, "Opacity");
-        var storyboard = new Storyboard();
-        storyboard.Children.Add(fade);
-        storyboard.Begin();
-    }
-
-    private async System.Threading.Tasks.Task ShowExportRelease(string releaseId)
-    {
-        var (settingsCurrent, settings) = await RunForCurrentHandle(NativeBae.GetSettings);
-        if (!settingsCurrent)
-        {
-            return;
-        }
-        var choices = new List<(string Label, BridgeExportSelection Selection)>
-        {
-            (Loc.Chrome("track.export.original"), OriginalExportSelection()),
-        };
-        choices.AddRange(settings.ExportPresets
-            .Where(preset => preset.AppliesToRelease)
-            .Select(preset => (preset.Name, PresetExportSelection(preset.Id))));
-
-        var formatPicker = new ComboBox
-        {
-            Header = Loc.Chrome("settings.export.default_release_format"),
-            MinWidth = 260,
-        };
-        var defaultIndex = -1;
-        for (var index = 0; index < choices.Count; index++)
-        {
-            var choice = choices[index];
-            if (ExportSelectionsEqual(choice.Selection, settings.DefaultReleaseExportSelection))
-            {
-                defaultIndex = index;
-            }
-            formatPicker.Items.Add(new ComboBoxItem
-            {
-                Content = choice.Label,
-                Tag = choice.Selection,
-            });
-        }
-        if (defaultIndex < 0)
-        {
-            StatusText.Text = Loc.Chrome("track.export.prepare_failed");
-            return;
-        }
-        formatPicker.SelectedIndex = defaultIndex;
-
-        var dialog = new ContentDialog
-        {
-            Title = Loc.Chrome("settings.export.label"),
-            Content = formatPicker,
-            PrimaryButtonText = Loc.Chrome("settings.export.label"),
-            CloseButtonText = Loc.Chrome("action.cancel"),
-            XamlRoot = Content.XamlRoot,
-        };
-        var result = await dialog.ShowAsync();
-        if (result != ContentDialogResult.Primary)
-        {
-            return;
-        }
-        if (formatPicker.SelectedItem is not ComboBoxItem selectedItem
-            || selectedItem.Tag is not BridgeExportSelection selection)
-        {
-            StatusText.Text = Loc.Chrome("track.export.prepare_failed");
-            return;
-        }
-
-        var picker = new global::Windows.Storage.Pickers.FolderPicker();
-        picker.FileTypeFilter.Add("*");
-        WinRT.Interop.InitializeWithWindow.Initialize(
-            picker, WinRT.Interop.WindowNative.GetWindowHandle(this));
-        var folder = await picker.PickSingleFolderAsync();
-        if (folder is null)
-        {
-            return;
-        }
-
-        var (exportCurrent, error) = await RunForCurrentHandle(
-            handle => NativeBae.ExportRelease(handle, releaseId, folder.Path, selection));
-        if (!exportCurrent)
-        {
-            return;
-        }
-        if (error is not null)
-        {
-            StatusText.Text = error;
-        }
-    }
-
-    private static BridgeExportSelection OriginalExportSelection() => new BridgeExportSelection.Original();
-
-    private static BridgeExportSelection PresetExportSelection(string presetId) =>
-        new BridgeExportSelection.Preset(presetId);
-
-    private static bool ExportSelectionsEqual(BridgeExportSelection a, BridgeExportSelection b) =>
-        (a, b) switch
-        {
-            (BridgeExportSelection.Original, BridgeExportSelection.Original) => true,
-            (BridgeExportSelection.Preset left, BridgeExportSelection.Preset right) => left.PresetId == right.PresetId,
-            _ => false,
-        };
-
-    private async System.Threading.Tasks.Task ConfirmDeleteRelease(string releaseId)
-    {
-        if (string.IsNullOrEmpty(releaseId))
-        {
-            return;
-        }
-
-        var confirm = new ContentDialog
-        {
-            Title = Loc.Chrome("album.delete.title"),
-            Content = Loc.Chrome("album.delete.body"),
-            PrimaryButtonText = Loc.Chrome("action.delete"),
-            CloseButtonText = Loc.Chrome("action.cancel"),
-            XamlRoot = Content.XamlRoot,
-        };
-        if (await confirm.ShowAsync() != ContentDialogResult.Primary)
-        {
-            return;
-        }
-
-        // On success an invalidation refreshes the grid.
-        var (current, error) = await RunForCurrentHandle(
-            handle => NativeBae.DeleteRelease(handle, releaseId));
-        if (!current)
-        {
-            return;
-        }
-        if (error is not null)
-        {
-            await ShowError(Loc.Chrome("album.delete.failed"), error);
-        }
-    }
-
-    private async System.Threading.Tasks.Task ShowEditMetadata(string releaseId)
-    {
-        if (string.IsNullOrEmpty(releaseId))
-        {
-            return;
-        }
-
-        var (current, seeded) = WithCurrentHandle(
-            handle => NativeBae.ReleaseEditSeed(handle, releaseId).Edit);
-        if (!current)
-        {
-            return;
-        }
-        if (seeded is null)
-        {
-            await ShowError(Loc.Chrome("album.edit.load_failed"));
-            return;
-        }
-
-        var form = new ReleaseEditForm(seeded, 520);
-
-        var dialog = new ContentDialog
-        {
-            Title = Loc.Chrome("album.edit.title"),
-            Content = new ScrollViewer { Content = form.Panel },
-            PrimaryButtonText = Loc.Chrome("action.save"),
-            SecondaryButtonText = Loc.Chrome("album.edit.reset"),
-            CloseButtonText = Loc.Chrome("action.cancel"),
-            XamlRoot = Content.XamlRoot,
-        };
-
-        // Shape + write happen in Rust; on a validation/write error keep the
-        // dialog open and show the reason instead of committing.
-        dialog.PrimaryButtonClick += (_, args) =>
-        {
-            var (editCurrent, error) = WithCurrentHandle(
-                handle => NativeBae.ApplyReleaseEdit(handle, releaseId, form.ReadBack()));
-            if (!editCurrent)
-            {
-                args.Cancel = true;
-                return;
-            }
-            if (error is not null)
-            {
-                form.ErrorText.Text = error;
-                form.ErrorText.Visibility = Visibility.Visible;
-                args.Cancel = true;
-            }
-        };
-
-        // Reset to Source discards the in-progress edits and re-seeds the form
-        // from the release's stored metadata source (its original identity)
-        // without writing the DB. Keep the dialog open regardless; a deferral
-        // holds it through the async re-projection so it can't close mid-await.
-        dialog.SecondaryButtonClick += async (_, args) =>
-        {
-            args.Cancel = true;
-            var deferral = args.GetDeferral();
-            try
-            {
-                var (resetCurrent, fresh) = await RunForCurrentHandle(
-                    handle => NativeBae.ResetMetadataToSource(handle, releaseId).Edit);
-                if (!resetCurrent)
-                {
-                    return;
-                }
-                if (fresh is null)
-                {
-                    form.ErrorText.Text = Loc.Chrome("album.edit.reset_failed");
-                    form.ErrorText.Visibility = Visibility.Visible;
-                    return;
-                }
-
-                form.ErrorText.Visibility = Visibility.Collapsed;
-                form.Seed(fresh);
-            }
-            finally
-            {
-                deferral.Complete();
-            }
-        };
-
-        await dialog.ShowAsync();
-    }
-
-    private async System.Threading.Tasks.Task ShowReidentify(string releaseId, string seedArtist, string seedAlbum)
-    {
-        if (string.IsNullOrEmpty(releaseId))
-        {
-            return;
-        }
-
-        var artistBox = new TextBox { Header = Loc.Chrome("search.field.artist"), Text = seedArtist };
-        var albumBox = new TextBox { Header = Loc.Chrome("search.field.album"), Text = seedAlbum };
-        var sourceBox = new ComboBox { Header = Loc.Chrome("search.field.source") };
-        sourceBox.Items.Add("discogs");
-        sourceBox.Items.Add("musicbrainz");
-        sourceBox.SelectedIndex = 0;
-        var searchButton = new Button { Content = Loc.Chrome("action.search") };
-
-        var resultsList = new ListView
-        {
-            SelectionMode = ListViewSelectionMode.Single,
-            MaxHeight = 280,
-        };
-        var status = new TextBlock
-        {
-            Foreground = new SolidColorBrush(Microsoft.UI.Colors.Salmon),
-            TextWrapping = TextWrapping.Wrap,
-            Visibility = Visibility.Collapsed,
-        };
-
-        var form = new StackPanel { Spacing = 8, Width = 420 };
-        form.Children.Add(artistBox);
-        form.Children.Add(albumBox);
-        form.Children.Add(sourceBox);
-        form.Children.Add(searchButton);
-        form.Children.Add(resultsList);
-        form.Children.Add(status);
-
-        var dialog = new ContentDialog
-        {
-            Title = Loc.Chrome("album.reidentify.title"),
-            Content = new ScrollViewer { Content = form },
-            PrimaryButtonText = Loc.Chrome("album.reidentify.confirm"),
-            CloseButtonText = Loc.Chrome("action.cancel"),
-            XamlRoot = Content.XamlRoot,
-            IsPrimaryButtonEnabled = false,
-        };
-
-        var candidates = new List<ReleaseCandidateChoice>();
-
-        // The generated bridge search and commit both block on network/DB work; run them off
-        // the UI thread so the dialog stays responsive.
-        searchButton.Click += async (_, _) =>
-        {
-            var source = (string)sourceBox.SelectedItem;
-            var artist = artistBox.Text;
-            var album = albumBox.Text;
-            searchButton.IsEnabled = false;
-            var (current, search) = await RunForCurrentHandle(
-                handle => NativeBae.SearchReleases(handle, source, artist, album));
-            searchButton.IsEnabled = true;
-            if (!current)
-            {
-                return;
-            }
-
-            if (search.Error is not null)
-            {
-                status.Text = search.Error;
-                status.Visibility = Visibility.Visible;
-                return;
-            }
-
-            candidates = search.Candidates ?? [];
-            resultsList.ItemsSource = candidates.Select(candidate => candidate.Summary).ToList();
-            status.Text = Loc.Chrome("search.no_matches");
-            status.Visibility = candidates.Count == 0 ? Visibility.Visible : Visibility.Collapsed;
-            dialog.IsPrimaryButtonEnabled = false;
-        };
-
-        resultsList.SelectionChanged += (_, _) =>
-        {
-            dialog.IsPrimaryButtonEnabled = resultsList.SelectedIndex >= 0;
-        };
-
-        dialog.PrimaryButtonClick += async (_, args) =>
-        {
-            var index = resultsList.SelectedIndex;
-            if (index < 0 || index >= candidates.Count)
-            {
-                args.Cancel = true;
-                return;
-            }
-
-            var chosen = candidates[index];
-            var deferral = args.GetDeferral();
-            var (current, error) = await RunForCurrentHandle(
-                handle => NativeBae.ReidentifyRelease(handle, releaseId, chosen.ReleaseId, chosen.Source));
-            if (!current)
-            {
-                args.Cancel = true;
-                deferral.Complete();
-                return;
-            }
-            if (error is not null)
-            {
-                status.Text = error;
-                status.Visibility = Visibility.Visible;
-                args.Cancel = true;
-            }
-
-            deferral.Complete();
-        };
-
-        await dialog.ShowAsync();
-    }
-
-    private async System.Threading.Tasks.Task ShowGallery(string releaseId)
-    {
-        var (current, images) = WithCurrentHandle(
-            handle => NativeBae.Gallery(handle, releaseId).Items);
-        if (!current)
-        {
-            return;
-        }
-        if (images is null)
-        {
-            StatusText.Text = Loc.Chrome("gallery.load_failed");
-            return;
-        }
-
-        if (images.Length == 0)
-        {
-            return;
-        }
-
-        var index = 0;
-        var image = new Image { Stretch = Stretch.Uniform, MinHeight = 360, MinWidth = 360 };
-        var label = new TextBlock { HorizontalAlignment = HorizontalAlignment.Center };
-        void Show()
-        {
-            var item = images[index];
-            var handle = CurrentHandleOrNull();
-            if (handle == null)
-            {
-                return;
-            }
-            image.Source = CoverImage.LoadGalleryBytes(handle, releaseId, item.Source);
-            label.Text = $"{item.Label} ({index + 1}/{images.Length})";
-        }
-        Show();
-
-        var prev = new Button { Content = "‹" };
-        var next = new Button { Content = "›" };
-        prev.Click += (_, _) => { index = (index - 1 + images.Length) % images.Length; Show(); };
-        next.Click += (_, _) => { index = (index + 1) % images.Length; Show(); };
-
-        var nav = new StackPanel
-        {
-            Orientation = Orientation.Horizontal,
-            HorizontalAlignment = HorizontalAlignment.Center,
-            Spacing = 12,
-        };
-        nav.Children.Add(prev);
-        nav.Children.Add(label);
-        nav.Children.Add(next);
-
-        var content = new StackPanel { Spacing = 8 };
-        content.Children.Add(image);
-        content.Children.Add(nav);
-
-        var dialog = new ContentDialog
-        {
-            Title = Loc.Chrome("gallery.title"),
-            Content = content,
-            CloseButtonText = Loc.Chrome("action.close"),
-            XamlRoot = Content.XamlRoot,
-        };
-        await dialog.ShowAsync();
-    }
-
-    /// <summary>
-    /// Pick a new cover for <paramref name="releaseId"/>: the release's own image
-    /// files plus remote candidates fetched from MusicBrainz / Discogs. Selecting
-    /// one writes it as the release's cover; the album grid refreshes via the
-    /// invalidation the change emits. Errors surface inside this dialog,
-    /// since the window banner is occluded by the modal.
-    /// </summary>
-    private async System.Threading.Tasks.Task ShowChangeCover(string albumId, string releaseId)
-    {
-        if (string.IsNullOrEmpty(albumId) || string.IsNullOrEmpty(releaseId))
-        {
-            return;
-        }
-
-        var (imagesCurrent, releaseImages) = WithCurrentHandle(
-            handle => NativeBae.GetReleaseImages(handle, releaseId).Images);
-        if (!imagesCurrent)
-        {
-            return;
-        }
-        if (releaseImages is null)
-        {
-            StatusText.Text = Loc.Chrome("cover.images_load_failed");
-            return;
-        }
-
-        var content = new StackPanel { Spacing = 8, MinWidth = 460 };
-
-        // Errors from a failed remote fetch or a failed change surface here; the
-        // window-level banner is hidden behind this modal dialog.
-        var statusText = new TextBlock
-        {
-            TextWrapping = TextWrapping.Wrap,
-            Foreground = new SolidColorBrush(Microsoft.UI.Colors.Salmon),
-            Visibility = Visibility.Collapsed,
-        };
-        content.Children.Add(statusText);
-
-        var dialog = new ContentDialog
-        {
-            Title = Loc.Chrome("cover.change_title"),
-            Content = new ScrollViewer { Content = content, MaxHeight = 520 },
-            CloseButtonText = Loc.Chrome("action.close"),
-            XamlRoot = Content.XamlRoot,
-        };
-
-        // Apply a selection off the UI thread (a remote cover downloads bytes),
-        // then close on success or show the error in place.
-        async System.Threading.Tasks.Task Apply(BridgeCoverSelection selection)
-        {
-            statusText.Visibility = Visibility.Collapsed;
-            var (current, error) = await RunForCurrentHandle(
-                handle => NativeBae.ChangeCover(handle, albumId, releaseId, selection));
-            if (!current)
-            {
-                return;
-            }
-            if (error is null)
-            {
-                dialog.Hide();
-            }
-            else
-            {
-                statusText.Text = error;
-                statusText.Visibility = Visibility.Visible;
-            }
-        }
-
-        // A thumbnail tile that applies the selection when clicked.
-        Button Tile(ImageSource? source, string caption, BridgeCoverSelection selection)
-        {
-            var button = CoverTile(source, caption);
-            button.Click += async (_, _) => await Apply(selection);
-            return button;
-        }
-
-        if (releaseImages.Length > 0)
-        {
-            content.Children.Add(new TextBlock { Text = Loc.Chrome("cover.release_files") });
-            var fileGrid = new VariableSizedWrapGrid
-            {
-                Orientation = Orientation.Horizontal,
-                ItemWidth = 140,
-                ItemHeight = 160,
-            };
-            foreach (var file in releaseImages)
-            {
-                var handle = CurrentHandleOrNull();
-                if (handle == null)
-                {
-                    return;
-                }
-                var source = CoverImage.LoadGalleryBytes(
-                    handle, releaseId, new BridgeGallerySource.ReleaseFile(file.Id));
-                var selection = new BridgeCoverSelection.ReleaseImage(file.Id);
-                fileGrid.Children.Add(Tile(source, file.OriginalFilename, selection));
-            }
-
-            content.Children.Add(fileGrid);
-        }
-
-        content.Children.Add(new TextBlock { Text = Loc.Chrome("cover.remote_sources") });
-        var loading = new StackPanel
-        {
-            Orientation = Orientation.Horizontal,
-            Spacing = 8,
-        };
-        loading.Children.Add(new ProgressRing { IsActive = true, Width = 20, Height = 20 });
-        loading.Children.Add(new TextBlock { Text = Loc.Chrome("cover.fetching") });
-        content.Children.Add(loading);
-
-        var remoteGrid = new VariableSizedWrapGrid
-        {
-            Orientation = Orientation.Horizontal,
-            ItemWidth = 140,
-            ItemHeight = 160,
-        };
-        content.Children.Add(remoteGrid);
-
-        // Fetch the remote candidates off the UI thread, then fill the grid on
-        // resume. The dialog opens immediately with the release files shown and a
-        // spinner where the remote covers will land.
-        async System.Threading.Tasks.Task LoadRemote()
-        {
-            var (current, covers) = await RunForCurrentHandle(
-                handle => NativeBae.FetchRemoteCovers(handle, releaseId).Covers);
-            if (!current)
-            {
-                return;
-            }
-            loading.Visibility = Visibility.Collapsed;
-            if (covers is null)
-            {
-                statusText.Text = Loc.Chrome("cover.fetch_failed");
-                statusText.Visibility = Visibility.Visible;
-                return;
-            }
-
-            try
-            {
-                if (covers.Length == 0)
-                {
-                    remoteGrid.Children.Add(new TextBlock { Text = Loc.Chrome("cover.none_remote") });
-                    return;
-                }
-
-                foreach (var cover in covers)
-                {
-                    var source = new BitmapImage(new Uri(NativeBae.RemoteCoverThumbnailUrl(cover)));
-                    var selection = NativeBae.RemoteCoverSelection(cover);
-                    remoteGrid.Children.Add(Tile(source, cover.Label, selection));
-                }
-            }
-            catch (Exception ex)
-            {
-                // Fire-and-forget: a malformed cover URL or unexpected payload must
-                // surface here, not as an unobserved task exception.
-                statusText.Text = Loc.Chrome("cover.show_failed", "detail", ex.Message);
-                statusText.Visibility = Visibility.Visible;
-            }
-        }
-
-        _ = LoadRemote();
-        await dialog.ShowAsync();
+        await _albumDetail.Show(album.Id);
     }
 
     private async void OnStorageClick(object sender, RoutedEventArgs e)
@@ -3248,7 +1993,7 @@ public sealed partial class MainWindow : Window
         void PopulateExportSelection(ComboBox combo, Settings settings, bool release)
         {
             combo.Items.Clear();
-            var original = OriginalExportSelection();
+            var original = ExportSelection.Original();
             var selected = release ? settings.DefaultReleaseExportSelection : settings.DefaultTrackExportSelection;
             combo.Items.Add(new ComboBoxItem
             {
@@ -3258,7 +2003,7 @@ public sealed partial class MainWindow : Window
             });
             foreach (var preset in settings.ExportPresets.Where(p => release ? p.AppliesToRelease : p.AppliesToTrack))
             {
-                var selection = PresetExportSelection(preset.Id);
+                var selection = ExportSelection.Preset(preset.Id);
                 combo.Items.Add(new ComboBoxItem
                 {
                     Content = preset.Name,
@@ -3269,7 +2014,7 @@ public sealed partial class MainWindow : Window
         }
 
         bool SameExportSelection(BridgeExportSelection a, BridgeExportSelection b) =>
-            ExportSelectionsEqual(a, b);
+            ExportSelection.Equal(a, b);
 
         string CodecLabel(BridgeExportPresetCodec codec) => codec switch
         {
